bluesky-feed/requirements.txt
atproto==0.0.59
peewee~=3.16.2
Flask~=2.3.2
python-dotenv~=1.0.0
google-generativeai
numpy